+++ CHANGELOG.md
@@ -29,6 +29,9 @@ Increment the:
 * [CODE HEALTH] Move remaining API test helpers into anonymous namespaces
   [#4301](https://github.com/open-telemetry/opentelemetry-cpp/pull/4301)
 
+* [BUG] Make SocketAddr string parsing safe and reject malformed addresses
+  [#4292](https://github.com/open-telemetry/opentelemetry-cpp/pull/4292)
+
 * [BUG] Fix use-after-free when an HTTP handler closes the connection
   [#4289](https://github.com/open-telemetry/opentelemetry-cpp/pull/4289)